refactor: 추상화 클래스 재정의에 따라 딜러, 플레이어의 코드 수정 및 패키지 이동
